Added support for multiple rows. Rows will automatically be added and removed when appropriate.
Need to account for increasing window size now. Keep keyboard at bottom. It also becomes apparent we'll need some sort of indicator (cursor) as to where the next operation will take effect. With automatic rows, it may not always be abundantly clear which character will be edited. But it's pretty good for now.