Added Wordle dictionary display when entering text.
This of course is unfiltered at the moment. The dictionary is populated on startup and displayed after a letter is entered. I'm leaving it shown all the time afterwards. Of course, it's a ton of options, and really only looks passable on computers. The next step is to better support mobile devices.