Showing sidebar instead of full results for small screens.
This required a wrapping element around the "main" content, with the wrapper adding the new sidebar element to the far right end.

The scrollbar was tricky -- in order to scroll only the main content, with the effect of an auto-sized viewport within the window, the 0-height trick had to be moved to the wrapper, while leaving the overflow-y setting at the main content. This must be preventing the window scrollbar from handling the overflow, allowing the main content to do so.

The CSS also hides the full results for screens narrower than 640px, showing the sidebar instead. This makes it functional at the PC level, as well.

I still need to add behavior to the sidebar to show the results on button tap.

Also added logo and docuement title.