Results flyout for small screens.
Initialized the results display for window size is narrow. Hides entry words and keyboard.

Also removed the flex-grow on the sidebar, as it was messing up the spacing. Auto-hide the flyout if the window grows to a better size. This uses an effect, which I'm not crazy about, since it duplicates the logic and magic number from the CSS. Will look into prettying this up more.