Added support for changing letter "status" by selection.
This required the introduction of a Letter class and LetterStatus enum. The status increments as the user touches it, and this status determines which style is applied at the Row level.

One thing I'm not exactly thrilled by is how all the data (and logic) apparently must be stored at the top level. Of course, sharing a single state is good, but anything that interacts with that data, any business logic, must exist at the same level, as well. There's assuredly more I need to learn about project structure in React.