Made possibilities variable stateless.
I realized I was causing double renders when words changed, as it would cause possibilities to change. Both were states, but if the words state changes, it will force a rerender anyway, so possibilities being a state was unnecessary. This takes the number of renders back down to once per change (twice in StrictMode).