Smarter yellow/green letter tracking.
Primary inclusions for this are skipping green if a green already exists in the column and preventing the addition of more letters if the total of green and yellow would exceed five, the latter being the most complex.

Computing total letters required the tracking of known letters (stored as a Word) and the tracking of mystery letters. Mystery letters follow the regular rules, but programmatically it can be pretty complex. I opted for recalculating these two data stores each render, as opposed to making them states, because it's less bug-prone and doesn't cause multiple rerenders each change.

The significant feature left to do is smart highlighting of other rows on letter click. Although not necessary now for filtering, it will really increase quality of life for the user and be pretty cool, to boot.