Added basic letter highlighting between rows.
This involved a number of changes, including using the new structuredClone() for a deep copy. Moved functions out of Word class, since structuredClone() doesn't support them during deep copy.

So, now the known and mystery letters are only computed when changing status. The algorithm first uses the preceding rows with the proposed change to determine if the change should be allowed. If not, we increment the status (to the remaining option) and try again. If that doesn't work, we undo. If the change is allowed, we re-compute the known and mystery letters, now including the changed row. Then we clear the following rows and re-populate based on the resulting letters.

This accounts for the difficult case of multiple letters being in any number of places, as well as various niche cases, such as when a word had a yellow in a previous column and still has the letter in that column, but where now the letter has been found in that word in a different column.

I had wanted to allow the current row to determine how to update the other rows, but that was causing many problems, so we're using everything up to that point and updating the rows following.

It's unclear if I actually have to do the brunt of this logic inside the setState(). It seems like not great design by React, but I may be misunderstanding things. In other words, I'm operating under the assumption that any modifications to the state (e.g. to recalc known and mystery letters) must occur inside the setState. I guess I could clone the state outside the setter and modify it there, but I'm not sure if there's any benefit to that extra cost.

A final thing I had to add was to the calculation of the known and mystery letters. Before, it wasn't counting green's in its number of occurrences. This was just a bug, necessary for multiple occurrences of the same letter.

The final thing I need to add before polishing (and fixing bugs) is adjusting the status when adding letters. This will be decidedly easier than this checkin.