Ensuring new letters have correct status, based on previous words.
Simply checks the known and mystery letters against any preceding rows to determine if a letter needs to start with a non-default status. Of course, if there's a green present already for that column, we use that, as well.

One case this doesn't account for is when a previously occurring letter occurs sooner in a following word. In this case, that letter will be highlighted. If then a following letter is indicated by the user as the known/green, the recently highlighted letter will not be unhighlighted. This will result in indicating two occurrences of that letter, one yellow and one green, which may not be the user's intent. I chose not to address this right now, since the fix for the user is simple and the fix in code would be less so.