Added help text to indicate that letters are interactive.
Text will only display if the user has at least two complete words and hasn't changed any status. Text fades in and out until a status changes. The fade animation is structured such that the text will start by fading in only, as opposed to fading in and out constantly, even when not being shown.