Replaced help text with groundwork for Help/About.
Help/About will gently flash when we detect the user may need help. This can be stopped by either setting a status or clicking the icon. In the future, the latter will open a model dialog with the Help/About info.

This overall seems more flexible and more professional than showing some simple text in the header. It also gives me a chance to put my info on the site.