Replaced icon and title with a logo.
This simplifies a bit of the code, and in addition to allowing for cleaner mouse selection, it just looks better. Of course, using an image is less flexible than text+SVG, but it's also more dependable.

Also reduced the size requirements for the help dialog. This will help it support more (mobile) devices.