Results are filtered by current words.
This included fleshing out the WordsParser to also include which letters are absent, which are limited to an exact number of occurrences, and which indeces can't have which letters. The exclusions don't include the known letters, though this might be needed for other consumers (or just ignore the known indeces). If a letter is not present, it could be absent entirely, or it could be telling us exactly how many occurrences are in the answer.

In Results, we run the parser (again) and iterate over all possible words, checking the following:
- if it has absent letters
- are the known letters present
- if exact number of a letter is known, are those present
- are the mystery letters present
- are the mystery letters at new columns

The only thing about the design that I'm not crazy about is running the WordsParser twice for each word change (if the results are visible). Since we just pass the words around, it needs to be recalculated in the Results component. The alternative would be to manage all five outputs from the method as a separate state. Though the current method is slower, at least it's clean and simple.